writing metaData.json; check on server not to overwrite existing file

+++ connections.routes.js
@@ -3,6 +3,25 @@ const router = express.Router()
 const fs = require('fs')
 const path = require("path")
 
+
+// write metaData.json on server boot
+
+let dir = fs.readdirSync(path.resolve(__dirname, "./public/data"), "utf8")
+let metaDataContents = { visibleGames: {}, invisibleGames: {} }
+dir.forEach(file => {
+   if (file.startsWith("game")) {
+      filedata = JSON.parse(fs.readFileSync(path.resolve(__dirname, "./public/data", file)))
+      metaDataContents.visibleGames[(filedata.title)] = file
+   }
+})
+// write to metadata file
+fs.writeFileSync(path.resolve(__dirname, "./public/data/metaData.json"), JSON.stringify(metaDataContents), (err) => {
+   if (err) throw err;
+   console.log('MetaData has been written');
+}, { flag: "w" })
+
+
+
 router.use(express.static(path.join(__dirname, 'public')));
 
 
@@ -36,7 +55,15 @@ router.post("/addGame", function (req, res) {
       let gameJSON = JSON.stringify(body)
       let metaData = JSON.parse(fs.readFileSync(path.resolve(__dirname, "./public/data/metaData.json"), "utf8"))
       let gameName = body.title
-      let gameFileName = "game_" + Object.keys(metaData.visibleGames).length + ".json"
+      //let gameFileName = "game_" + Object.keys(metaData.visibleGames).length + ".json"
+      let gameFileName = "game_" + gameName + "_" + body.author + ".json"
+
+      // checks if filename already exists
+      let existsAlready = fs.readdirSync(path.resolve(__dirname, "./public/data")).includes(gameFileName)
+      if (existsAlready) {
+         res.sendStatus(404)
+         return
+      }
 
       // create file
       fs.writeFile(path.resolve(__dirname, "./public/data/", gameFileName), gameJSON, (err) => {
@@ -54,10 +81,11 @@ router.post("/addGame", function (req, res) {
 
 
       res.sendStatus(200)
+      return
    }
-   else {
-      res.sendStatus(404)
-   }
+
+   res.sendStatus(404)
+
 
 })
